New feature: Graphical map editor for maps extension
The map editor allows for exporting and importing maps in the form of wikicode,
it supports all map overlays and it's metadata as well as the majority of
the global parameters available for maps.

The editor resides in a special page (Special:MapEditor) and handles
all overlay types (lines,rectangles,polygons,circles,markers etc)

The editor is NOT a WYSIWYG editor, there will be visual differences in
the maps representation by the native implementation and the editor.

The editor does NOT support using named locations instead of coordinates.

** Fixed the issues you pointed out.
** as for the minColors library checking for jquery, this is a third party
library, and i am not comfortable with editing this for a small issue like this
** Added i18n

***fixed i18n bug

Change-Id: I01cf625398debaa0380fc4963c6c8de60ed36c1e