improved gui on searchmarkers
in order to save space the descriptive text for the input box for filtering markers
is placed as a value of the input field instead of a text on the left hand side.

when the input field is focused, the text is removed (if not changed) and the user can type their
query to filter on.

when the input field looses focus, the text is reverted back to the original state if no text was
entered into the field.

Change-Id: I0f5212d7957c0054aa723595404364ff6a614f5d