added support for standard parameter "link" in maps format
default behaviour was to always add a link to the map marker bubble, however
now if you set link=none as a parameter to the ask query, it will no longer
enforce the link in the marker bubble.

I believe this is more in line with what is expected out of the standard
parameters in semantic media wiki.

also added support for standard parameter "headers" (plain,show,hide)

a value of plain will show properties key -> val but it wont format the key as
a html link.

a value of hide, will hide all the properties

Change-Id: I80529bc1f5cdc01237866cdfbc11db2f15067c28