Added a new parameter (activeicon)) to semantic maps.
If you have an ask query with format map where the result gives a set of pages
with coordinate data. Each result page also has its own map with the same
ask query.

The activeicon parameter will then show a different marker icon for the page
you are on in the map. So if your query is gives a list of countries, and you
are currently on the page Turkey, then the map of the countries will have a
single marker in Turkey as a different icon than the rest.

Example usage:

{{#ask: [[Category:Country]]
|?Coordinate
|format=map
|activeicon=File:Blue-dot.png
}}

or

{{#ask: [[Category:Country]]
|?Coordinate
|format=Map
|activeicon=/some/path/to/blue-dot.png
}}

Change-Id: Ieda8985cd03c256d7ec287f99ade3d2d0ce02c5a