Fixed special case when searchbox is enabled. There is a special case when the searchbox is enabled where one can have markers and other map objects beneath the search input box.
This decreases usability.

However by panning the map -30 pixels on the y axis before the map is loaded,
this prevents that. And it also seems to prevent other map opbjects to be
pushed outside the map if they were within the 30px range which was panned out
of the map. This is probably due to the event type used when issuing the panBy
method. As it probably kicks in before the map calculates its own zoom and
center based on the bounding box.

Change-Id: I90fb6973298432631ee4de0e9f4ffb24742a6842