Re-introduces custom image base layers for OpenLayers
This has initially been committed to gerrit.wikimedia.org on Oct 18, 2012 02:47
See https://gerrit.wikimedia.org/r/#/c/28430/

This feature has been part of some old version of Maps already but got broken a long time ago. This fixes all issues and re-introduces the original feature plus various improvements to Maps. Custom layers have to be defined in a Layer namespace. All layers will be saved in the database, therefore running update.php is required for this to work.
This feature is still rather experimental in this commit.