[DNM] Use latest version of Leaflet
Not getting this to work properly...
Been trying for an hour to get the resources loaded, but my wiki
is not allowing access to files in vendor/ it seems. Might be that
it is configured badly. I don't want to spend more time on MW shit,
though would be awesome if someone figures out what the issue is.