Make tests compatible with PHPUnit 4 and 6
Use the PHPUnit4And6Compat trait from MediaWiki core to make PHPUnit 4
methods work even when being run with PHPUnit 6. If it's being run with
an older MediaWiki core version, define a empty trait in the bootstrap
to avoid fatal errors. Since the lack of the trait means we're running
with PHPUnit 4, everything should work out.