Remove use of deprecated wgStyleVersion
I couldn't find any use of egMapsStyleVersion, but to keep this commit simple,
I kept that variable and only removed the use of wgStyleVersion. Also, in case
something depends on its format I've substituted it with the number 1.

This way any use of it will still work and still invalidate cache if the Maps_VERSION
constant is changed.